jit: split common lowering and optimize self recursion
